requirements.txt
bech32==1.2.0
discord.py==2.3.2
httpx>=0.27.0
PyYAML>=6.0.1
python-dotenv==1.0.1